fix: collapse组件部分抽象到component
