+++ scss/_properties.scss
@@ -537,6 +537,12 @@
   --Collapse-content-fontSize: var(--body-size);
   --Collapse-content-fontWeight: var(--body-weight);
 
+  --Collapse2-border-color: #{px2rem(1px)} solid #EBEDF5;
+  --Collapse2-header-padding: #{px2rem(14px)};
+  --Collapse2-header-bg-color: #F6F7FA;
+  --Collapse2-header-bg-hover-color: #F3F9FE;
+  --Collapse2-header-bg-disabled-color: #F6F7FA;
+
   --Crud-toolbar-gap: var(--gap-xs);
   --Crud-toolbar-height: #{px2rem(30px)};
   --Crud-toolbar-lineHeight: var(--lineHeightBase);
